refactor: move telemetry work to more general location
